# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cybernix Nexus API Specification",
    "version": "2.0.0",
    "description": "Official OpenAPI 3.1 specification for Cybernix Nexus — NSEC Multi-Platform Competitive Programming Hub. Provides endpoints for student rating synchronization, real-time leaderboards, peer problem editorials, contest scheduling, and background cron scraping.",
    "contact": {
      "name": "Phoenix Tech Club (NSEC)",
      "url": "https://github.com/cybernixNexus"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://cybernixnexus.nsec.ac.in",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "summary": "Register New Student Account",
        "description": "Creates a new user and student profile with academic info and coding platform handles, validating roll number uniqueness and triggering background platform stats sync.",
        "tags": [
          "Authentication & Onboarding"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Account created successfully"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    },
                    "hasStudentProfile": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required credentials or invalid data"
          },
          "409": {
            "description": "Email or Roll Number already registered in college directory"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/cron/sync": {
      "get": {
        "summary": "Execute Scheduled Platform Scraping Cron",
        "description": "Secured cron worker for syncing all eligible student coding metrics and recording daily snapshots.",
        "tags": [
          "Cron & Background Tasks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sync execution summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "syncedCount": {
                      "type": "integer"
                    },
                    "eligibleCount": {
                      "type": "integer"
                    },
                    "skippedCount": {
                      "type": "integer"
                    },
                    "totalCount": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid CRON_SECRET bearer token"
          }
        }
      }
    },
...
    "/api/editorials": {
      "get": {
        "summary": "List Peer Problem Tutorials & Solutions",
        "description": "Fetches community tutorials filtered by platform, difficulty, tags, or search query with author profiles and upvote counts.",
        "tags": [
          "Peer Editorials"
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by platform (Codeforces, LeetCode, etc.)"
          },
          {
            "name": "difficulty",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by difficulty (Easy, Medium, Hard)"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for title or tags"
          }
        ],
        "responses": {
          "200": {
            "description": "List of peer tutorials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "editorials": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Editorial"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Publish Peer Problem Solution",
        "description": "Creates a new peer problem tutorial with code snippets, markdown explanations, tags, and problem source URLs.",
        "tags": [
          "Peer Editorials"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEditorialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tutorial published successfully"
          },
          "401": {
            "description": "Authentication required to publish tutorials"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token for background cron operations (CRON_SECRET)."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth persistent session cookie."
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "rollNumber",
          "department",
          "graduationYear"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Satyaki Das"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "satyaki@nsec.ac.in"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "secretpassword123"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://example.com/avatar.jpg"
          },
          "rollNumber": {
            "type": "string",
            "example": "10800121001"
          },
          "department": {
            "type": "string",
            "enum": [
              "CSE",
              "IT",
              "ECE",
              "AI&DS",
              "EE",
              "ME"
            ],
            "example": "CSE"
          },
          "graduationYear": {
            "type": "integer",
            "example": 2026
          },
          "github": {
            "type": "string",
            "nullable": true,
            "example": "octocat"
          },
          "linkedin": {
            "type": "string",
            "nullable": true,
            "example": "satyaki-das"
          },
          "leetcode": {
            "type": "string",
            "nullable": true,
            "example": "neal_wu"
          },
          "codeforces": {
            "type": "string",
            "nullable": true,
            "example": "tourist"
          },
          "gfg": {
            "type": "string",
            "nullable": true,
            "example": "gfg_student"
          },
          "codechef": {
            "type": "string",
            "nullable": true,
            "example": "chef_student"
          }
        }
      },
...
      "Editorial": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "problemUrl": {
            "type": "string",
            "format": "uri"
          },
          "platform": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ]
          },
          "authorName": {
            "type": "string"
          },
          "authorDept": {
            "type": "string"
          },
          "authorLevel": {
            "type": "integer"
          },
          "authorTier": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "codeSnippet": {
            "type": "string",
            "nullable": true
          },
          "codeLanguage": {
            "type": "string",
            "nullable": true
          },
          "likesCount": {
            "type": "integer"
          },
          "commentsCount": {
            "type": "integer"
          },
          "isLikedByMe": {
            "type": "boolean"
          },
          "publishedAt": {
            "type": "string"
          }
        }
      },
      "CreateEditorialRequest": {
        "type": "object",
        "required": [
          "title",
          "problemUrl",
          "platform",
          "difficulty",
          "tags",
          "summary",
          "content"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "problemUrl": {
            "type": "string",
            "format": "uri"
          },
          "platform": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "codeSnippet": {
            "type": "string"
          },
          "codeLanguage": {
            "type": "string",
            "default": "cpp"
          }
        }
      },
...
    }
  }
}